try using `_api_template.py` to generate different APIs
